requirements.txt
fastapi
uvicorn
python-multipart
sqlalchemy
pymysql
pandas
scipy
PyJWT
passlib
bcrypt
